+++ application/shared/view/timemanager/includes/navbar-admin.php
@@ -12,7 +12,8 @@
     <div class="navbar-collapse collapse">
       <ul class="nav navbar-nav">
         <li {home_active}><a href="{timemanager_root}main">Home</a></li>
-        <li {jobs_active}><a href="{timemanager_root}jobs">Jobs - Quotes</a></li>
+        <li {jobs_quoting_active}><a href="{timemanager_root}jobs">Jobs - Quotes</a></li>
+        <li {jobs_tracking_active}><a href="{timemanager_root}jobs/tracking">Jobs - Tracking</a></li>
         <li {about_active}><a href="{timemanager_root}about">About</a></li>
         <li {settings_active}><a href="{timemanager_root}settings">Settings</a></li>
         <li><a href="{timemanager_root}logout">Logout</a></l>